requirements.txt
tensorflow-gpu
pillow
h5py
